Introduce SPI to OpenAiClient
