+++ src/main/java/dev/ai4j/openai4j/OpenAiClient.java
@@ -9,118 +9,66 @@ import dev.ai4j.openai4j.embedding.EmbeddingResponse;
 import dev.ai4j.openai4j.moderation.ModerationRequest;
 import dev.ai4j.openai4j.moderation.ModerationResponse;
 import dev.ai4j.openai4j.moderation.ModerationResult;
-import okhttp3.Cache;
-import okhttp3.OkHttpClient;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
-
-import java.io.IOException;
+import dev.ai4j.openai4j.spi.OpenAiClientBuilderFactory;
+import dev.ai4j.openai4j.spi.ServiceHelper;
 import java.net.InetSocketAddress;
 import java.net.Proxy;
 import java.time.Duration;
 import java.util.List;
 
-import static dev.ai4j.openai4j.Json.GSON;
+public abstract class OpenAiClient {
 
-public class OpenAiClient {
+    public abstract SyncOrAsyncOrStreaming<CompletionResponse> completion(CompletionRequest request);
 
-    private static final Logger log = LoggerFactory.getLogger(OpenAiClient.class);
+    @Experimental
+    public abstract SyncOrAsyncOrStreaming<String> completion(String prompt);
 
-    private final String baseUrl;
-    private final String apiVersion;
-    private final OkHttpClient okHttpClient;
-    private final OpenAiApi openAiApi;
-    private final boolean logStreamingResponses;
+    public abstract SyncOrAsyncOrStreaming<ChatCompletionResponse> chatCompletion(ChatCompletionRequest request);
 
-    public OpenAiClient(String apiKey) {
-        this(builder().openAiApiKey(apiKey));
+    @Experimental
+    public abstract SyncOrAsyncOrStreaming<String> chatCompletion(String userMessage);
+
+    public abstract SyncOrAsync<EmbeddingResponse> embedding(EmbeddingRequest request);
+
+    @Experimental
+    public abstract SyncOrAsync<List<Float>> embedding(String input);
+
+    public abstract SyncOrAsync<ModerationResponse> moderation(ModerationRequest request);
+
+    @Experimental
+    public abstract SyncOrAsync<ModerationResult> moderation(String input);
+
+    public abstract void shutdown();
+
+    @SuppressWarnings("rawtypes")
+    public static OpenAiClient.Builder builder() {
+        for (OpenAiClientBuilderFactory factory : ServiceHelper.loadFactories(OpenAiClientBuilderFactory.class)) {
+            return factory.get();
+        }
+        // fallback to the default
+        return DefaultOpenAiClient.builder();
+
     }
 
-    private OpenAiClient(Builder serviceBuilder) {
 
-        this.baseUrl = serviceBuilder.baseUrl;
-        this.apiVersion = serviceBuilder.apiVersion;
+    @SuppressWarnings("unchecked")
+    public abstract static class Builder<T extends OpenAiClient, B extends Builder<T, B>> {
 
-        OkHttpClient.Builder okHttpClientBuilder = new OkHttpClient.Builder()
-                .callTimeout(serviceBuilder.callTimeout)
-                .connectTimeout(serviceBuilder.connectTimeout)
-                .readTimeout(serviceBuilder.readTimeout)
-                .writeTimeout(serviceBuilder.writeTimeout);
+        public String baseUrl = "https://api.openai.com/v1/";
+        public String apiVersion;
+        public String openAiApiKey;
+        public String azureApiKey;
+        public Duration callTimeout = Duration.ofSeconds(60);
+        public Duration connectTimeout = Duration.ofSeconds(60);
+        public Duration readTimeout = Duration.ofSeconds(60);
+        public Duration writeTimeout = Duration.ofSeconds(60);
+        public Proxy proxy;
+        public boolean logRequests;
+        public boolean logResponses;
+        public boolean logStreamingResponses;
 
-        if (serviceBuilder.openAiApiKey == null && serviceBuilder.azureApiKey == null) {
-            throw new IllegalArgumentException("openAiApiKey OR azureApiKey must be defined");
-        }
-        if (serviceBuilder.openAiApiKey != null && serviceBuilder.azureApiKey != null) {
-            throw new IllegalArgumentException("openAiApiKey AND azureApiKey cannot both be defined at the same time");
-        }
-        if (serviceBuilder.openAiApiKey != null) {
-            okHttpClientBuilder.addInterceptor(new AuthorizationHeaderInjector(serviceBuilder.openAiApiKey));
-        } else {
-            okHttpClientBuilder.addInterceptor(new ApiKeyHeaderInjector(serviceBuilder.azureApiKey));
-        }
 
-        if (serviceBuilder.proxy != null) {
-            okHttpClientBuilder.proxy(serviceBuilder.proxy);
-        }
-
-        if (serviceBuilder.logRequests) {
-            okHttpClientBuilder.addInterceptor(new RequestLoggingInterceptor());
-        }
-        if (serviceBuilder.logResponses) {
-            okHttpClientBuilder.addInterceptor(new ResponseLoggingInterceptor());
-        }
-        this.logStreamingResponses = serviceBuilder.logStreamingResponses;
-
-        this.okHttpClient = okHttpClientBuilder.build();
-
-        Retrofit retrofit = new Retrofit.Builder()
-                .baseUrl(serviceBuilder.baseUrl)
-                .client(okHttpClient)
-                .addConverterFactory(GsonConverterFactory.create(GSON))
-                .build();
-
-        this.openAiApi = retrofit.create(OpenAiApi.class);
-    }
-
-    public void shutdown() {
-
-        okHttpClient.dispatcher().executorService().shutdown();
-
-        okHttpClient.connectionPool().evictAll();
-
-        Cache cache = okHttpClient.cache();
-        if (cache != null) {
-            try {
-                cache.close();
-            } catch (IOException e) {
-                log.error("Failed to close cache", e);
-            }
-        }
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    public static class Builder {
-
-        private String baseUrl = "https://api.openai.com/v1/";
-        private String apiVersion;
-        private String openAiApiKey;
-        private String azureApiKey;
-        private Duration callTimeout = Duration.ofSeconds(60);
-        private Duration connectTimeout = Duration.ofSeconds(60);
-        private Duration readTimeout = Duration.ofSeconds(60);
-        private Duration writeTimeout = Duration.ofSeconds(60);
-        private Proxy proxy;
-        private boolean logRequests;
-        private boolean logResponses;
-        private boolean logStreamingResponses;
-
-        private Builder() {
-        }
+        public abstract T build();
 
         /**
          * @param baseUrl Base URL of OpenAI API.
@@ -128,21 +76,21 @@ public class OpenAiClient {
          *                For Azure OpenAI: "https://{resource-name}.openai.azure.com/openai/deployments/{deployment-id}/"
          * @return builder
          */
-        public Builder baseUrl(String baseUrl) {
+        public B baseUrl(String baseUrl) {
             if (baseUrl == null || baseUrl.trim().isEmpty()) {
                 throw new IllegalArgumentException("baseUrl cannot be null or empty");
             }
             this.baseUrl = baseUrl.endsWith("/") ? baseUrl : baseUrl + "/";
-            return this;
+            return (B) this;
         }
 
         /**
          * @param apiVersion Version of the API in the YYYY-MM-DD format. Applicable only for Azure OpenAI.
          * @return builder
          */
-        public Builder apiVersion(String apiVersion) {
+        public B apiVersion(String apiVersion) {
             this.apiVersion = apiVersion;
-            return this;
+            return (B) this;
         }
 
         /**
@@ -150,12 +98,12 @@ public class OpenAiClient {
          *                     Will be injected in HTTP headers like this: "Authorization: Bearer ${openAiApiKey}"
          * @return builder
          */
-        public Builder openAiApiKey(String openAiApiKey) {
+        public B openAiApiKey(String openAiApiKey) {
             if (openAiApiKey == null || openAiApiKey.trim().isEmpty()) {
                 throw new IllegalArgumentException("openAiApiKey cannot be null or empty. API keys can be generated here: https://platform.openai.com/account/api-keys");
             }
             this.openAiApiKey = openAiApiKey;
-            return this;
+            return (B) this;
         }
 
         /**
@@ -163,221 +111,92 @@ public class OpenAiClient {
          *                    Will be injected in HTTP headers like this: "api-key: ${azureApiKey}"
          * @return builder
          */
-        public Builder azureApiKey(String azureApiKey) {
+        public B azureApiKey(String azureApiKey) {
             if (azureApiKey == null || azureApiKey.trim().isEmpty()) {
                 throw new IllegalArgumentException("azureApiKey cannot be null or empty");
             }
             this.azureApiKey = azureApiKey;
-            return this;
+            return (B) this;
         }
 
-        public Builder callTimeout(Duration callTimeout) {
+        public B callTimeout(Duration callTimeout) {
             if (callTimeout == null) {
                 throw new IllegalArgumentException("callTimeout cannot be null");
             }
             this.callTimeout = callTimeout;
-            return this;
+            return (B) this;
         }
 
-        public Builder connectTimeout(Duration connectTimeout) {
+        public B connectTimeout(Duration connectTimeout) {
             if (connectTimeout == null) {
                 throw new IllegalArgumentException("connectTimeout cannot be null");
             }
             this.connectTimeout = connectTimeout;
-            return this;
+            return (B) this;
         }
 
-        public Builder readTimeout(Duration readTimeout) {
+        public B readTimeout(Duration readTimeout) {
             if (readTimeout == null) {
                 throw new IllegalArgumentException("readTimeout cannot be null");
             }
             this.readTimeout = readTimeout;
-            return this;
+            return (B) this;
         }
 
-        public Builder writeTimeout(Duration writeTimeout) {
+        public B writeTimeout(Duration writeTimeout) {
             if (writeTimeout == null) {
                 throw new IllegalArgumentException("writeTimeout cannot be null");
             }
             this.writeTimeout = writeTimeout;
-            return this;
+            return (B) this;
         }
 
-        public Builder proxy(Proxy.Type type, String ip, int port) {
+        public B proxy(Proxy.Type type, String ip, int port) {
             this.proxy = new Proxy(type, new InetSocketAddress(ip, port));
-            return this;
+            return (B) this;
         }
 
-        public Builder proxy(Proxy proxy) {
+        public B proxy(Proxy proxy) {
             this.proxy = proxy;
-            return this;
+            return (B) this;
         }
 
-        public Builder logRequests() {
+        public B logRequests() {
             return logRequests(true);
         }
 
-        public Builder logRequests(Boolean logRequests) {
+        public B logRequests(Boolean logRequests) {
             if (logRequests == null) {
                 logRequests = false;
             }
             this.logRequests = logRequests;
-            return this;
+            return (B) this;
         }
 
-        public Builder logResponses() {
+        public B logResponses() {
             return logResponses(true);
         }
 
-        public Builder logResponses(Boolean logResponses) {
+        public B logResponses(Boolean logResponses) {
             if (logResponses == null) {
                 logResponses = false;
             }
             this.logResponses = logResponses;
-            return this;
+            return (B) this;
         }
 
-        public Builder logStreamingResponses() {
+        public B logStreamingResponses() {
             return logStreamingResponses(true);
         }
 
-        public Builder logStreamingResponses(Boolean logStreamingResponses) {
+        public B logStreamingResponses(Boolean logStreamingResponses) {
             if (logStreamingResponses == null) {
                 logStreamingResponses = false;
             }
             this.logStreamingResponses = logStreamingResponses;
-            return this;
+            return (B) this;
         }
 
-        public OpenAiClient build() {
-            return new OpenAiClient(this);
-        }
-    }
 
-    public SyncOrAsyncOrStreaming<CompletionResponse> completion(CompletionRequest request) {
-
-        CompletionRequest syncRequest = CompletionRequest.builder()
-                .from(request)
-                .stream(null)
-                .build();
-
-        return new RequestExecutor<>(
-                openAiApi.completions(syncRequest, apiVersion),
-                (r) -> r,
-                okHttpClient,
-                formatUrl("completions"),
-                () -> CompletionRequest.builder().from(request).stream(true).build(),
-                CompletionResponse.class,
-                (r) -> r,
-                logStreamingResponses
-        );
-    }
-
-    @Experimental
-    public SyncOrAsyncOrStreaming<String> completion(String prompt) {
-
-        CompletionRequest request = CompletionRequest.builder()
-                .prompt(prompt)
-                .build();
-
-        CompletionRequest syncRequest = CompletionRequest.builder()
-                .from(request)
-                .stream(null)
-                .build();
-
-        return new RequestExecutor<>(
-                openAiApi.completions(syncRequest, apiVersion),
-                CompletionResponse::text,
-                okHttpClient,
-                formatUrl("completions"),
-                () -> CompletionRequest.builder().from(request).stream(true).build(),
-                CompletionResponse.class,
-                CompletionResponse::text,
-                logStreamingResponses
-        );
-    }
-
-    public SyncOrAsyncOrStreaming<ChatCompletionResponse> chatCompletion(ChatCompletionRequest request) {
-
-        ChatCompletionRequest syncRequest = ChatCompletionRequest.builder()
-                .from(request)
-                .stream(null)
-                .build();
-
-        return new RequestExecutor<>(
-                openAiApi.chatCompletions(syncRequest, apiVersion),
-                (r) -> r,
-                okHttpClient,
-                formatUrl("chat/completions"),
-                () -> ChatCompletionRequest.builder().from(request).stream(true).build(),
-                ChatCompletionResponse.class,
-                (r) -> r,
-                logStreamingResponses
-        );
-    }
-
-    @Experimental
-    public SyncOrAsyncOrStreaming<String> chatCompletion(String userMessage) {
-
-        ChatCompletionRequest request = ChatCompletionRequest.builder()
-                .addUserMessage(userMessage)
-                .build();
-
-        ChatCompletionRequest syncRequest = ChatCompletionRequest.builder()
-                .from(request)
-                .stream(null)
-                .build();
-
-        return new RequestExecutor<>(
-                openAiApi.chatCompletions(syncRequest, apiVersion),
-                ChatCompletionResponse::content,
-                okHttpClient,
-                formatUrl("chat/completions"),
-                () -> ChatCompletionRequest.builder().from(request).stream(true).build(),
-                ChatCompletionResponse.class,
-                (r) -> r.choices().get(0).delta().content(),
-                logStreamingResponses
-        );
-    }
-
-    public SyncOrAsync<EmbeddingResponse> embedding(EmbeddingRequest request) {
-
-        return new RequestExecutor<>(openAiApi.embeddings(request, apiVersion), (r) -> r);
-    }
-
-    @Experimental
-    public SyncOrAsync<List<Float>> embedding(String input) {
-
-        EmbeddingRequest request = EmbeddingRequest.builder()
-                .input(input)
-                .build();
-
-        return new RequestExecutor<>(openAiApi.embeddings(request, apiVersion), EmbeddingResponse::embedding);
-    }
-
-    public SyncOrAsync<ModerationResponse> moderation(ModerationRequest request) {
-
-        return new RequestExecutor<>(openAiApi.moderations(request, apiVersion), (r) -> r);
-    }
-
-    @Experimental
-    public SyncOrAsync<ModerationResult> moderation(String input) {
-
-        ModerationRequest request = ModerationRequest.builder()
-                .input(input)
-                .build();
-
-        return new RequestExecutor<>(openAiApi.moderations(request, apiVersion), (r) -> r.results().get(0));
-    }
-
-    private String formatUrl(String endpoint) {
-        return baseUrl + endpoint + apiVersionQueryParam();
-    }
-
-    private String apiVersionQueryParam() {
-        if (apiVersion == null || apiVersion.trim().isEmpty()) {
-            return "";
-        }
-        return "?api-version=" + apiVersion;
     }
 }
